Created MilestoneService and refactored MilestonesController to use the new service. #2299
