+++ src/main/webapp/wise5/classroomMonitor/milestones/milestones.html
@@ -2,7 +2,7 @@
     <div class="l-constrained">
         <div layout="row" layout-align="center center" layout-align-gt-xs="start center" layout-wrap>
             <!-- TODO: make into MilestoneItem component -->
-            <md-card ng-repeat="milestone in milestonesController.projectAchievements | orderBy:'params.targetDateTemp'"
+            <md-card ng-repeat="milestone in milestonesController.milestones | orderBy:'params.targetDateTemp'"
                      class="milestone md-button" md-ink-ripple
                      ng-click="milestonesController.showMilestoneDetails(milestone, $event)"
                      aria-label="{{ ::'VIEW_MILESTONE_DETAILS' | translate:{ name: milestone.name } }}">
@@ -19,7 +19,7 @@
                             <md-icon ng-if="milestone.percentageCompleted === 100" class="success md-48"> check </md-icon>
                         </div>
                     </div>
-                    <p class="md-body-1 text-secondary">{{ milestone.numberOfStudentsCompleted }}/{{ milestonesController.numberOfStudentsInRun }} teams completed</p>
+                    <p class="md-body-1 text-secondary">{{ milestone.numberOfStudentsCompleted }}/{{ milestone.numberOfStudentsInRun }} teams completed</p>
                     <p ng-if="milestone.isReportAvailable" class="md-body-2 info">~ {{ ::'reportAvailable' | translate }} ~</p>
                     <div class="md-body-2" ng-if="!milestone.isReportAvailable"
                          ng-class="{ 'warn': milestonesController.isBeforeDay(milestone.params.targetDate, milestone.percentageCompleted),
@@ -28,7 +28,7 @@
                     </div>
                 </md-card-content>
             </md-card>
-            <!--<md-card class="milestone milestone--add md-button primary" md-ink-ripple
+            <!-- <md-card class="milestone milestone--add md-button primary" md-ink-ripple
                      ng-click="milestonesController.editMilestone(null, $event)"
                      aria-label="{{ ::'ADD_MILESTONE' | translate }}">
                 <md-card-content layout="column" layout-align="center center">
@@ -37,7 +37,7 @@
                         <md-icon class="md-48 md-primary"> add </md-icon>
                     </div>
                 </md-card-content>
-            </md-card>-->
+            </md-card> -->
         </div>
     </div>
 </div>
