+++ src/main/webapp/wise5/classroomMonitor/milestones/milestonesController.ts
@@ -1,42 +1,19 @@
 'use strict';
 
 import AchievementService from '../../services/achievementService';
-import AnnotationService from '../../services/annotationService';
-import ConfigService from '../../services/configService';
-import ClassroomMonitorProjectService from '../classroomMonitorProjectService';
-import TeacherDataService from '../../services/teacherDataService';
-import UtilService from '../../services/utilService';
+import MilestoneService from '../../services/milestoneService';
 import * as angular from 'angular';
 
 class MilestonesController {
   $translate: any;
-  /*
-   * Arrays used to temporarily store milestone display values. We add
-   * fields to the milestone objects but we don't want to save those
-   * fields when we save the milestones to the server. We remove the
-   * fields from the milestones and then save the milestones to the
-   * server. After we save the milestones, we add the fields back into
-   * the milestones.
-   */
-  itemsTemporaryStorage: any[] = [];
-  numberOfStudentsCompletedStorage: any[] = [];
-  numberOfStudentsInRun: number;
-  percentageCompletedStorage: any[] = [];
-  periodId: any;
-  projectAchievements: any[];
-  workgroupIds: any[];
-  workgroupsStorage: any[] = [];
+  milestones: any[];
   static $inject = [
     '$filter',
     '$mdDialog',
     '$rootScope',
     '$scope',
     'AchievementService',
-    'AnnotationService',
-    'ConfigService',
-    'ProjectService',
-    'TeacherDataService',
-    'UtilService',
+    'MilestoneService',
     'moment'
   ];
   constructor(
@@ -45,75 +22,57 @@ class MilestonesController {
     private $rootScope: any,
     private $scope: any,
     private AchievementService: AchievementService,
-    private AnnotationService: AnnotationService,
-    private ConfigService: ConfigService,
-    private ProjectService: ClassroomMonitorProjectService,
-    private TeacherDataService: TeacherDataService,
-    private UtilService: UtilService,
+    private MilestoneService: MilestoneService,
     private moment: any
   ) {
     this.$translate = this.$filter('translate');
-    this.periodId = this.TeacherDataService.getCurrentPeriod().periodId;
-    this.setWorkgroupsInCurrentPeriod();
-    this.loadProjectAchievements();
+    this.loadProjectMilestones();
 
     this.$rootScope.$on('newStudentAchievement', (event, args) => {
       if (args) {
         const studentAchievement = args.studentAchievement;
         this.AchievementService.addOrUpdateStudentAchievement(studentAchievement);
-        this.updateMilestoneCompletion(studentAchievement.achievementId);
+        this.updateMilestoneStatus(studentAchievement.id);
       }
     });
 
     this.$scope.$on('currentPeriodChanged', (event, args) => {
-      this.periodId = args.currentPeriod.periodId;
-
-      // update the completion status for all the project projectAchievements
-      for (let projectAchievement of this.projectAchievements) {
-        this.setWorkgroupsInCurrentPeriod();
-        this.updateMilestoneCompletion(projectAchievement.id);
+      for (let milestone of this.milestones) {
+        this.updateMilestoneStatus(milestone.id);
       }
     });
 
     this.$scope.$on('annotationReceived', (event, args) => {
-      for (const projectAchievement of this.projectAchievements) {
+      for (const milestone of this.milestones) {
         if (
-          projectAchievement.nodeId === args.annotation.nodeId &&
-          projectAchievement.componentId === args.annotation.componentId
+          milestone.nodeId === args.annotation.nodeId &&
+          milestone.componentId === args.annotation.componentId
         ) {
-          this.updateMilestoneCompletion(projectAchievement.id);
+          this.updateMilestoneStatus(milestone.id);
         }
       }
     });
   }
 
-  /**
-   * Load the projectAchievements and perform additional calculations
-   */
-  loadProjectAchievements() {
-    const projectAchievements = this.ProjectService.getAchievements();
-    if (projectAchievements.isEnabled) {
-      this.projectAchievements = projectAchievements.items;
-      for (let projectAchievement of this.projectAchievements) {
-        this.updateMilestoneCompletion(projectAchievement.id);
+  loadProjectMilestones() {
+    this.milestones = this.MilestoneService.getProjectMilestones();
+    for (let milestone of this.milestones) {
+      milestone = this.MilestoneService.getProjectMilestoneStatus(milestone.id);
+    }
+  }
 
-        // get all the activities and steps in the project
-        projectAchievement.items = this.UtilService.makeCopyOfJSONObject(
-          this.ProjectService.idToOrder
-        );
-        if (projectAchievement.params != null && projectAchievement.params.nodeIds != null) {
-          /*
-           * loop through all the node ids that are required
-           * to be completed for this project achievement
-           */
-          for (let nodeId of projectAchievement.params.nodeIds) {
-            if (projectAchievement.items[nodeId] != null) {
-              projectAchievement.items[nodeId].checked = true;
-            }
-          }
-        }
+  updateMilestoneStatus(milestoneId) {
+    let milestone = this.getProjectMilestoneById(milestoneId);
+    milestone = this.MilestoneService.getProjectMilestoneStatus(milestoneId);
+  }
+
+  getProjectMilestoneById(milestoneId) {
+    for (let milestone of this.milestones) {
+      if (milestone.id === milestoneId) {
+        return milestone;
       }
     }
+    return {};
   }
 
   /**
@@ -146,663 +105,6 @@ class MilestonesController {
     return result;
   }
 
-  /**
-   * Create a new milestone
-   * @return a milestone object
-   */
-  createMilestone() {
-    let projectAchievements = this.ProjectService.getAchievementItems();
-    if (projectAchievements != null) {
-      // get the time of tomorrow at 3pm
-      const tomorrow = this.moment()
-        .add('days', 1)
-        .hours(23)
-        .minutes(11)
-        .seconds(59);
-      return {
-        id: this.AchievementService.getAvailableAchievementId(),
-        name: '',
-        description: '',
-        type: 'milestone',
-        params: {
-          nodeIds: [],
-          targetDate: tomorrow.valueOf()
-        },
-        icon: {
-          image: ''
-        },
-        items: this.UtilService.makeCopyOfJSONObject(this.ProjectService.idToOrder),
-        isVisible: true
-      };
-    }
-    return null;
-  }
-
-  /**
-   * Delete a milestone
-   * @param milestone the milestone to delete
-   */
-  deleteMilestone(milestone, $event) {
-    if (milestone) {
-      const title = milestone.name;
-      const label = this.$translate('DELETE_MILESTONE');
-      const msg = this.$translate('DELETE_MILESTONE_CONFIRM', {
-        name: milestone.name
-      });
-      const yes = this.$translate('YES');
-      const cancel = this.$translate('CANCEL');
-
-      const confirm = this.$mdDialog
-        .confirm()
-        .title(title)
-        .textContent(msg)
-        .ariaLabel(label)
-        .targetEvent($event)
-        .ok(yes)
-        .cancel(cancel);
-
-      this.$mdDialog.show(confirm).then(
-        () => {
-          let projectAchievements = this.projectAchievements;
-          let index = -1;
-          for (let i = 0; i < projectAchievements.length; i++) {
-            if (projectAchievements[i].id === milestone.id) {
-              index = i;
-              break;
-            }
-          }
-
-          if (index > -1) {
-            this.projectAchievements.splice(index, 1);
-            this.saveProject();
-          }
-        },
-        () => {}
-      );
-    }
-  }
-
-  saveMilestone(milestone) {
-    let index = -1;
-    for (let i = 0; i < this.projectAchievements.length; i++) {
-      if (this.projectAchievements[i].id === milestone.id) {
-        index = i;
-        this.projectAchievements[i] = milestone;
-        break;
-      }
-    }
-    if (index < 0) {
-      let projectAchievements = this.ProjectService.getAchievementItems();
-      if (projectAchievements && milestone) {
-        projectAchievements.push(milestone);
-      }
-    }
-    this.saveProject();
-    this.loadProjectAchievements();
-  }
-
-  /**
-   * Remove the temporary fields from the milestone objects and store
-   * them in temporary storage arrays so that we can load the fields back
-   * in later
-   */
-  clearTempFields() {
-    /*
-     * these array will store the temporary fields. the index of the arrays corresponds to the
-     * index of the project achievement. for example the percentageCompletedStorage value for
-     * the first project project achievement will be stored in
-     * this.percentageCompletedStorage[0]. the percentageCompletedStorage value for the second
-     * project project achievement will be stored in this.percentageCompletedStorage[1].
-     */
-    this.itemsTemporaryStorage = [];
-    this.workgroupsStorage = [];
-    this.numberOfStudentsCompletedStorage = [];
-    this.percentageCompletedStorage = [];
-
-    for (let projectAchievement of this.projectAchievements) {
-      // save the field values in the temporary storage arrays
-      this.workgroupsStorage.push(projectAchievement.workgroups);
-      this.numberOfStudentsCompletedStorage.push(projectAchievement.numberOfStudentsCompleted);
-      this.percentageCompletedStorage.push(projectAchievement.percentageCompleted);
-
-      // delete the field from the projectAchievement
-      delete projectAchievement.items;
-      delete projectAchievement.workgroups;
-      delete projectAchievement.numberOfStudentsCompleted;
-      delete projectAchievement.percentageCompleted;
-    }
-  }
-
-  /**
-   * Restore the temporary fields into the achievement objects
-   */
-  restoreTempFields() {
-    for (let a = 0; a < this.projectAchievements.length; a++) {
-      const projectAchievement = this.projectAchievements[a];
-      // set the fields back into the achievement object
-      projectAchievement.items = this.itemsTemporaryStorage[a];
-      projectAchievement.workgroups = this.workgroupsStorage[a];
-      projectAchievement.numberOfStudentsCompleted = this.numberOfStudentsCompletedStorage[a];
-      projectAchievement.percentageCompleted = this.percentageCompletedStorage[a];
-    }
-    this.itemsTemporaryStorage = [];
-    this.workgroupsStorage = [];
-    this.numberOfStudentsCompletedStorage = [];
-    this.percentageCompletedStorage = [];
-  }
-
-  /**
-   * Save the project to the server
-   */
-  saveProject() {
-    this.clearTempFields();
-    this.ProjectService.saveProject();
-    this.restoreTempFields();
-  }
-
-  /**
-   * Get the user names for a workgroup id
-   * @param workgroupId the workgroup id
-   * @return the user names in the workgroup
-   */
-  getDisplayUsernamesByWorkgroupId(workgroupId) {
-    return this.ConfigService.getDisplayUsernamesByWorkgroupId(workgroupId);
-  }
-
-  setWorkgroupsInCurrentPeriod() {
-    const workgroupIdsInRun = this.ConfigService.getClassmateWorkgroupIds();
-    this.workgroupIds = [];
-
-    // filter out workgroups not in the current period
-    for (let i = 0; i < workgroupIdsInRun.length; i++) {
-      const currentId = workgroupIdsInRun[i];
-      const currentPeriodId = this.ConfigService.getPeriodIdByWorkgroupId(currentId);
-
-      if (this.periodId === -1 || currentPeriodId === this.periodId) {
-        this.workgroupIds.push(currentId);
-      }
-    }
-    this.numberOfStudentsInRun = this.workgroupIds.length;
-  }
-
-  /**
-   * Update the student completion information for this milestone
-   * @param achievementId the achievement id to update
-   */
-  updateMilestoneCompletion(achievementId) {
-    const projectAchievement = this.getProjectAchievementById(achievementId);
-    const achievementIdToStudentAchievements = this.AchievementService.getAchievementIdToStudentAchievementsMappings();
-    const studentAchievements = achievementIdToStudentAchievements[projectAchievement.id];
-    const workgroupIdsCompleted = [];
-    const achievementTimes = [];
-    const workgroupIdsNotCompleted = [];
-
-    for (let studentAchievement of studentAchievements) {
-      const currentWorkgroupId = studentAchievement.workgroupId;
-      // check if workgroup is in current period
-      if (this.workgroupIds.indexOf(currentWorkgroupId) > -1) {
-        workgroupIdsCompleted.push(currentWorkgroupId);
-        achievementTimes.push(studentAchievement.achievementTime);
-      }
-    }
-
-    for (let workgroupId of this.workgroupIds) {
-      if (workgroupIdsCompleted.indexOf(workgroupId) === -1) {
-        workgroupIdsNotCompleted.push(workgroupId);
-      }
-    }
-
-    projectAchievement.workgroups = [];
-
-    for (let c = 0; c < workgroupIdsCompleted.length; c++) {
-      const workgroupId = workgroupIdsCompleted[c];
-      const achievementTime = achievementTimes[c];
-      const workgroupObject = {
-        workgroupId: workgroupId,
-        displayNames: this.getDisplayUsernamesByWorkgroupId(workgroupId),
-        achievementTime: achievementTime,
-        completed: true
-      };
-      projectAchievement.workgroups.push(workgroupObject);
-    }
-
-    /*
-     * loop through all the workgroups that have not
-     * completed the achievement
-     */
-    for (let workgroupId of workgroupIdsNotCompleted) {
-      const workgroupObject = {
-        workgroupId: workgroupId,
-        displayNames: this.getDisplayUsernamesByWorkgroupId(workgroupId),
-        achievementTime: null,
-        completed: false
-      };
-      projectAchievement.workgroups.push(workgroupObject);
-    }
-
-    projectAchievement.numberOfStudentsCompleted = workgroupIdsCompleted.length;
-    projectAchievement.percentageCompleted =
-      Math.round((100 * projectAchievement.numberOfStudentsCompleted) / this.numberOfStudentsInRun);
-
-    if (projectAchievement.type === 'milestoneReport') {
-      if (this.isCompletionReached(projectAchievement)) {
-        const report = this.generateReport(projectAchievement);
-        this.setReportAvailable(projectAchievement, true);
-        projectAchievement.generatedReport = report.content ? report.content : null;
-        projectAchievement.recommendations = report.recommendations ? report.recommendations : null;
-        projectAchievement.nodeId = report.nodeId;
-        projectAchievement.componentId = report.componentId;
-      } else {
-        this.setReportAvailable(projectAchievement, false);
-      }
-    }
-  }
-
-  isCompletionReached(projectAchievement) {
-    return (
-      projectAchievement.percentageCompleted >= projectAchievement.satisfyMinPercentage &&
-      projectAchievement.numberOfStudentsCompleted >= projectAchievement.satisfyMinNumWorkgroups
-    );
-  }
-
-  setReportAvailable(projectAchievement, reportAvailable) {
-    projectAchievement.isReportAvailable = reportAvailable;
-  }
-
-  generateReport(projectAchievement) {
-    const referencedComponents = this.getSatisfyCriteriaReferencedComponents(projectAchievement);
-    const aggregateAutoScores = {};
-    let nodeId = null;
-    let componentId = null;
-    const referencedComponentValues: any[] = Object.values(referencedComponents);
-    for (const referencedComponent of referencedComponentValues) {
-      nodeId = referencedComponent.nodeId;
-      componentId = referencedComponent.componentId;
-      aggregateAutoScores[componentId] = this.calculateAggregateAutoScores(
-        nodeId,
-        componentId,
-        this.periodId,
-        projectAchievement.report
-      );
-    }
-    const template = this.chooseTemplate(projectAchievement.report.templates, aggregateAutoScores);
-    let content = template.content ? template.content : '';
-    if (content) {
-      content = this.processMilestoneGraphsAndData(content, aggregateAutoScores);
-    }
-    const recommendations = template.recommendations ? template.recommendations : '';
-    return {
-      content: content,
-      recommendations: recommendations,
-      nodeId: nodeId,
-      componentId: componentId
-    };
-  }
-
-  processMilestoneGraphsAndData(content, aggregateAutoScores) {
-    for (const componentAggregate of Object.values(aggregateAutoScores)) {
-      let subScoreIndex = 0;
-      for (let [subScoreId, aggregateData] of Object.entries(componentAggregate)) {
-        let index = 0;
-        if (subScoreId !== 'ki') {
-          subScoreIndex++;
-          index = subScoreIndex;
-        }
-        const data = JSON.stringify(aggregateData).replace(/\"/g, "'");
-        const graphRegex = new RegExp(`milestone-report-graph{1,} id="(${subScoreId})"`, 'g');
-        content = content.replace(graphRegex, `$& data=\"${data}\"`);
-        const dataRegex = new RegExp(`milestone-report-data{1,} score-id="(${subScoreId})"`, 'g');
-        content = content.replace(dataRegex, `$& data=\"${data}\"`);
-      }
-    }
-    return content;
-  }
-
-  getSatisfyCriteriaReferencedComponents(projectAchievement) {
-    const components = {};
-    const templates = projectAchievement.report.templates;
-    for (const template of templates) {
-      for (const satisfyCriterion of template.satisfyCriteria) {
-        const nodeId = satisfyCriterion.nodeId;
-        const componentId = satisfyCriterion.componentId;
-        const component = {
-          nodeId: nodeId,
-          componentId: componentId
-        };
-        components[nodeId + '_' + componentId] = component;
-      }
-    }
-    return components;
-  }
-
-  calculateAggregateAutoScores(nodeId, componentId, periodId, reportSettings) {
-    const aggregate = {};
-    const scoreAnnotations = this.AnnotationService.getAllLatestScoreAnnotations(
-      nodeId,
-      componentId,
-      periodId
-    );
-    for (const scoreAnnotation of scoreAnnotations) {
-      if (scoreAnnotation.type === 'autoScore') {
-        this.addDataToAggregate(aggregate, scoreAnnotation, reportSettings);
-      } else {
-        const autoScoreAnnotation = this.AnnotationService.getLatestScoreAnnotation(
-          nodeId,
-          componentId,
-          scoreAnnotation.toWorkgroupId,
-          'autoScore'
-        );
-        if (autoScoreAnnotation) {
-          const mergedAnnotation = this.mergeAutoScoreAndTeacherScore(
-            autoScoreAnnotation,
-            scoreAnnotation,
-            reportSettings
-          );
-          this.addDataToAggregate(aggregate, mergedAnnotation, reportSettings);
-        }
-      }
-    }
-    return aggregate;
-  }
-
-  mergeAutoScoreAndTeacherScore(autoScoreAnnotation, teacherScoreAnnotation, reportSettings) {
-    if (autoScoreAnnotation.data.scores) {
-      for (const subScore of autoScoreAnnotation.data.scores) {
-        if (subScore.id === 'ki') {
-          subScore.score = this.adjustKIScore(teacherScoreAnnotation.data.value, reportSettings);
-        }
-      }
-    }
-    return autoScoreAnnotation;
-  }
-
-  adjustKIScore(scoreValue, reportSettings) {
-    const teacherScore = Math.round(scoreValue);
-    const kiScoreBounds = this.getKIScoreBounds(reportSettings);
-    let score = teacherScore;
-    if (teacherScore > kiScoreBounds.max) {
-      score = kiScoreBounds.max;
-    }
-    if (teacherScore < kiScoreBounds.min) {
-      score = kiScoreBounds.min;
-    }
-    return score;
-  }
-
-  getKIScoreBounds(reportSettings) {
-    const bounds = {
-      min: 1,
-      max: 5
-    };
-    if (reportSettings.customScoreValues && reportSettings.customScoreValues['ki']) {
-      bounds.min = Math.min(...reportSettings.customScoreValues['ki']);
-      bounds.max = Math.max(...reportSettings.customScoreValues['ki']);
-    }
-    return bounds;
-  }
-
-  addDataToAggregate(aggregate, annotation, reportSettings) {
-    for (const subScore of annotation.data.scores) {
-      if (aggregate[subScore.id] == null) {
-        aggregate[subScore.id] = this.setupAggregateSubScore(subScore.id, reportSettings);
-      }
-      const subScoreVal = subScore.score;
-      if (aggregate[subScore.id].counts[subScoreVal] > -1) {
-        aggregate[subScore.id].counts[subScoreVal]++;
-        aggregate[subScore.id].scoreSum += subScoreVal;
-        aggregate[subScore.id].scoreCount++;
-        aggregate[subScore.id].average =
-          aggregate[subScore.id].scoreSum / aggregate[subScore.id].scoreCount;
-      }
-    }
-    return aggregate;
-  }
-
-  setupAggregateSubScore(subScoreId, reportSettings) {
-    let counts = {};
-    if (reportSettings.customScoreValues && reportSettings.customScoreValues[subScoreId]) {
-      counts = this.getCustomScoreValueCounts(reportSettings.customScoreValues[subScoreId]);
-    } else {
-      counts = this.getPossibleScoreValueCounts(subScoreId);
-    }
-    return {
-      scoreSum: 0,
-      scoreCount: 0,
-      counts: counts,
-      average: 0
-    };
-  }
-
-  getCustomScoreValueCounts(scoreValues) {
-    let counts = {};
-    for (const value of scoreValues) {
-      counts[value] = 0;
-    }
-    return counts;
-  }
-
-  getPossibleScoreValueCounts(subScoreId) {
-    if (subScoreId === 'ki') {
-      return {
-        1: 0,
-        2: 0,
-        3: 0,
-        4: 0,
-        5: 0
-      };
-    } else {
-      return {
-        1: 0,
-        2: 0,
-        3: 0
-      };
-    }
-  }
-
-  chooseTemplate(templates, aggregateAutoScores) {
-    for (let template of templates) {
-      if (this.isTemplateMatch(template, aggregateAutoScores)) {
-        return template;
-      }
-    }
-    return {
-      content: null
-    };
-  }
-
-  isTemplateMatch(template, aggregateAutoScores) {
-    const matchedCriteria = [];
-    for (const satisfyCriterion of template.satisfyCriteria) {
-      if (this.isTemplateCriterionSatisfied(satisfyCriterion, aggregateAutoScores)) {
-        matchedCriteria.push(satisfyCriterion);
-      }
-    }
-    if (template.satisfyConditional === 'all') {
-      return matchedCriteria.length === template.satisfyCriteria.length;
-    } else if (template.satisfyConditional === 'any') {
-      return matchedCriteria.length > 0;
-    }
-  }
-
-  isTemplateCriterionSatisfied(satisfyCriterion, aggregateAutoScores) {
-    if (satisfyCriterion.function === 'percentOfScoresGreaterThan') {
-      return this.isPercentOfScoresGreaterThan(satisfyCriterion, aggregateAutoScores);
-    } else if (satisfyCriterion.function === 'percentOfScoresGreaterThanOrEqualTo') {
-      return this.isPercentOfScoresGreaterThanOrEqualTo(satisfyCriterion, aggregateAutoScores);
-    } else if (satisfyCriterion.function === 'percentOfScoresLessThan') {
-      return this.isPercentOfScoresLessThan(satisfyCriterion, aggregateAutoScores);
-    } else if (satisfyCriterion.function === 'percentOfScoresLessThanOrEqualTo') {
-      return this.isPercentOfScoresLessThanOrEqualTo(satisfyCriterion, aggregateAutoScores);
-    } else if (satisfyCriterion.function === 'percentOfScoresEqualTo') {
-      return this.isPercentOfScoresEqualTo(satisfyCriterion, aggregateAutoScores);
-    } else if (satisfyCriterion.function === 'percentOfScoresNotEqualTo') {
-      return this.isPercentOfScoresNotEqualTo(satisfyCriterion, aggregateAutoScores);
-    } else if (satisfyCriterion.function === 'default') {
-      return true;
-    }
-  }
-
-  isPercentOfScoresGreaterThan(satisfyCriterion, aggregateAutoScores) {
-    const aggregateData = this.getAggregateData(satisfyCriterion, aggregateAutoScores);
-    const possibleScores = this.getPossibleScores(aggregateData);
-    const sum = this.getGreaterThanSum(satisfyCriterion, aggregateData, possibleScores);
-    return this.isPercentThresholdSatisfied(satisfyCriterion, aggregateData, sum);
-  }
-
-  getGreaterThanSum(satisfyCriterion, aggregateData, possibleScores) {
-    let sum = 0;
-    for (const possibleScore of possibleScores) {
-      if (possibleScore > satisfyCriterion.value) {
-        sum += aggregateData.counts[possibleScore];
-      }
-    }
-    return sum;
-  }
-
-  isPercentOfScoresGreaterThanOrEqualTo(satisfyCriterion, aggregateAutoScores) {
-    const aggregateData = this.getAggregateData(satisfyCriterion, aggregateAutoScores);
-    const possibleScores = this.getPossibleScores(aggregateData);
-    const sum = this.getGreaterThanOrEqualToSum(satisfyCriterion, aggregateData, possibleScores);
-    return this.isPercentThresholdSatisfied(satisfyCriterion, aggregateData, sum);
-  }
-
-  getGreaterThanOrEqualToSum(satisfyCriterion, aggregateData, possibleScores) {
-    let sum = 0;
-    for (const possibleScore of possibleScores) {
-      if (possibleScore >= satisfyCriterion.value) {
-        sum += aggregateData.counts[possibleScore];
-      }
-    }
-    return sum;
-  }
-
-  isPercentOfScoresLessThan(satisfyCriterion, aggregateAutoScores) {
-    const aggregateData = this.getAggregateData(satisfyCriterion, aggregateAutoScores);
-    const possibleScores = this.getPossibleScores(aggregateData);
-    const sum = this.getLessThanSum(satisfyCriterion, aggregateData, possibleScores);
-    return this.isPercentThresholdSatisfied(satisfyCriterion, aggregateData, sum);
-  }
-
-  getLessThanSum(satisfyCriterion, aggregateData, possibleScores) {
-    let sum = 0;
-    for (const possibleScore of possibleScores) {
-      if (possibleScore < satisfyCriterion.value) {
-        sum += aggregateData.counts[possibleScore];
-      }
-    }
-    return sum;
-  }
-
-  isPercentOfScoresLessThanOrEqualTo(satisfyCriterion, aggregateAutoScores) {
-    const aggregateData = this.getAggregateData(satisfyCriterion, aggregateAutoScores);
-    const possibleScores = this.getPossibleScores(aggregateData);
-    const sum = this.getLessThanOrEqualToSum(satisfyCriterion, aggregateData, possibleScores);
-    return this.isPercentThresholdSatisfied(satisfyCriterion, aggregateData, sum);
-  }
-
-  getLessThanOrEqualToSum(satisfyCriterion, aggregateData, possibleScores) {
-    let sum = 0;
-    for (const possibleScore of possibleScores) {
-      if (possibleScore <= satisfyCriterion.value) {
-        sum += aggregateData.counts[possibleScore];
-      }
-    }
-    return sum;
-  }
-
-  isPercentOfScoresEqualTo(satisfyCriterion, aggregateAutoScores) {
-    const aggregateData = this.getAggregateData(satisfyCriterion, aggregateAutoScores);
-    const possibleScores = this.getPossibleScores(aggregateData);
-    const sum = this.getEqualToSum(satisfyCriterion, aggregateData, possibleScores);
-    return this.isPercentThresholdSatisfied(satisfyCriterion, aggregateData, sum);
-  }
-
-  getEqualToSum(satisfyCriterion, aggregateData, possibleScores) {
-    let sum = 0;
-    for (const possibleScore of possibleScores) {
-      if (possibleScore === satisfyCriterion.value) {
-        sum += aggregateData.counts[possibleScore];
-      }
-    }
-    return sum;
-  }
-
-  isPercentOfScoresNotEqualTo(satisfyCriterion, aggregateAutoScores) {
-    const aggregateData = this.getAggregateData(satisfyCriterion, aggregateAutoScores);
-    const possibleScores = this.getPossibleScores(aggregateData);
-    const sum = this.getNotEqualToSum(satisfyCriterion, aggregateData, possibleScores);
-    return this.isPercentThresholdSatisfied(satisfyCriterion, aggregateData, sum);
-  }
-
-  getNotEqualToSum(satisfyCriterion, aggregateData, possibleScores) {
-    let sum = 0;
-    for (const possibleScore of possibleScores) {
-      if (possibleScore !== satisfyCriterion.value) {
-        sum += aggregateData.counts[possibleScore];
-      }
-    }
-    return sum;
-  }
-
-  getAggregateData(satisfyCriterion, aggregateAutoScores) {
-    const component = aggregateAutoScores[satisfyCriterion.componentId];
-    return component[satisfyCriterion.targetVariable];
-  }
-
-  getPossibleScores(aggregateData) {
-    return Object.keys(aggregateData.counts)
-      .map(Number)
-      .sort();
-  }
-
-  isPercentThresholdSatisfied(satisfyCriterion, aggregateData, sum) {
-    const percentOfScores = (100 * sum) / aggregateData.scoreCount;
-    return percentOfScores >= satisfyCriterion.percentThreshold;
-  }
-
-  getProjectAchievementById(achievementId) {
-    for (let projectAchievement of this.projectAchievements) {
-      if (projectAchievement.id === achievementId) {
-        return projectAchievement;
-      }
-    }
-    return {};
-  }
-
-  generateName(scoreName) {
-    if (scoreName === 'ki') {
-      return `name="KI Score"`;
-    } else if (scoreName === 'science') {
-      return `name="Science Score"`;
-    } else if (scoreName === 'engineering') {
-      return `name="Engineering Score"`;
-    }
-  }
-
-  generateCategories(scoreName) {
-    if (scoreName === 'ki') {
-      return `categories="['1', '2', '3', '4', '5']"`;
-    } else {
-      return `categories="['0', '1', '2', '3']"`;
-    }
-  }
-
-  generateData(scoreName) {
-    if (scoreName === 'ki') {
-      return `name="KI Score"`;
-    } else if (scoreName === 'science') {
-      return `name="Science Score"`;
-    } else if (scoreName === 'engineering') {
-      return `name="Engineering Score"`;
-    }
-  }
-
-  /**
-   * Open a dialog with the milestone details (list with workgroups statuses
-   * for the given milestone)
-   * @param milestone the milestone object to show
-   * @param $event the event that triggered the function call
-   */
   showMilestoneDetails(milestone, $event) {
     const title = this.$translate('MILESTONE_DETAILS_TITLE', {
       name: milestone.name
@@ -826,7 +128,7 @@ class MilestonesController {
             <span flex></span>
             <md-button class="md-primary"
                        ng-click="edit()"
-                       ng-if="!milestone.type === 'milestoneReport'"
+                       ng-if="milestone.type != 'milestoneReport'"
                        aria-label="{{ ::'EDIT' | translate }}">
               {{ ::'EDIT' | translate }}
             </md-button>
@@ -934,17 +236,12 @@ class MilestonesController {
       );
   }
 
-  /**
-   * Open a dialog to edit milestone details (or create a new one)
-   * @param milestone the milestone object to show
-   * @param $event the event that triggered the function call
-   */
   editMilestone(milestone, $event) {
     let editMode = milestone ? true : false;
     let title = editMode ? this.$translate('EDIT_MILESTONE') : this.$translate('ADD_MILESTONE');
 
     if (!editMode) {
-      milestone = this.createMilestone();
+      milestone = this.MilestoneService.createMilestone();
     }
 
     let template = `<md-dialog class="dialog--wide">
@@ -960,6 +257,11 @@ class MilestonesController {
             <md-button ng-click="close()"
                        aria-label="{{ ::'CANCEL' | translate }}">
               {{ ::'CANCEL' | translate }}
+            </md-button>
+            <md-button class="md-warn"
+                       ng-click="delete()"
+                       aria-label="{{ ::'DELETE' | translate }}">
+              {{ ::'DELETE' | translate }}
             </md-button>
             <md-button class="md-primary"
                        ng-click="save()"
@@ -1018,6 +320,14 @@ class MilestonesController {
               }
             };
 
+            $scope.delete = function() {
+              $mdDialog.hide({
+                milestone: $scope.milestone,
+                delete: true,
+                $event: $scope.$event
+              });
+            };
+
             $scope.onChange = function(milestone, valid) {
               $scope.milestone = milestone;
               $scope.valid = valid;
@@ -1028,13 +338,52 @@ class MilestonesController {
       .then(
         data => {
           if (data) {
-            if (data.milestone && data.save) {
-              this.saveMilestone(data.milestone);
+            if (data.milestone) {
+              if (data.save) {
+                this.saveMilestone(data.milestone);
+              }
+              if (data.delete) {
+                this.deleteMilestone(data.milestone, $event);
+              }
             }
           }
         },
         () => {}
       );
+  }
+
+  deleteMilestone(milestone, $event) {
+    if (milestone) {
+      const title = milestone.name;
+      const label = this.$translate('DELETE_MILESTONE');
+      const msg = this.$translate('DELETE_MILESTONE_CONFIRM', {
+        name: milestone.name
+      });
+      const yes = this.$translate('YES');
+      const cancel = this.$translate('CANCEL');
+
+      const confirm = this.$mdDialog
+        .confirm()
+        .title(title)
+        .textContent(msg)
+        .ariaLabel(label)
+        .targetEvent($event)
+        .ok(yes)
+        .cancel(cancel);
+
+      this.$mdDialog.show(confirm).then(
+        () => {
+          this.MilestoneService.deleteMilestone(milestone);
+          this.loadProjectMilestones();
+        },
+        () => {}
+      );
+    }
+  }
+
+  saveMilestone(milestone) {
+    this.MilestoneService.saveMilestone(milestone);
+    this.loadProjectMilestones();
   }
 }
